decouple VendorDataFactory
